# src/main/resources/openapi/user-service-openapi.yaml
openapi: 3.1.0
info:
  title: User Service API - Image Gallery Microservices
  description: User Management Microservice for Image Gallery application.
  contact:
    name: cannibal-kektor
    url: https://github.com/cannibal-kektor
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearer-token: []
- internal-service: []
tags:
- name: User Management API
  description: Provides API for registering new users and retrieving information about
    existing users.
paths:
  /api/users/register:
    post:
      tags:
      - User Management API
      summary: Register new user
      description: Creates a new user in the system.
      operationId: register
      requestBody:
        description: User registration data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
        required: true
      responses:
        "200":
          description: Actual user dto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
  /api/users/username/{username}:
    get:
      tags:
      - User Management API
      summary: Get user by username
      description: Retrieves user information by their unique username
      operationId: getByUsername
      parameters:
      - name: username
        in: path
        description: Unique username (login)
        required: true
        schema:
          type: string
          example: alex_white
          maxLength: 30
          minLength: 3
          pattern: "^[a-z0-9_-]{3,30}$"
      responses:
        "200":
          description: Actual user dto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
      security:
      - internal-service: []
  /api/users/id/{id}:
    get:
      tags:
      - User Management API
      summary: Get user by ID
      description: Retrieves user information by their id
      operationId: getById
      parameters:
      - name: id
        in: path
        description: Unique user identifier
        required: true
        schema:
          type: integer
          format: int64
          example: 12345
          minimum: 1
      responses:
        "200":
          description: Actual user dto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
      security:
      - internal-service: []
  /api/users/current:
    get:
      tags:
      - User Management API
      summary: Get current authenticated user
      description: Retrieves information about the currently authenticated user.
      operationId: getCurrent
      responses:
        "200":
          description: Actual user dto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/ProblemDetailResponse"
        "401":
          $ref: "#/components/responses/ProblemDetailResponse"
        "403":
          $ref: "#/components/responses/ProblemDetailResponse"
        "404":
          $ref: "#/components/responses/ProblemDetailResponse"
        "429":
          $ref: "#/components/responses/ProblemDetailResponse"
        "500":
          $ref: "#/components/responses/ProblemDetailResponse"
        "503":
          $ref: "#/components/responses/ProblemDetailResponse"
      security:
      - bearer-token: []
components:
  schemas:
    RegistrationRequest:
      type: object
      description: Request model for registering a new user in the Image Gallery system
      properties:
        username:
          type: string
          description: Unique username (login)
          example: alex_white
          maxLength: 30
          minLength: 3
          pattern: "^[a-z0-9_-]{3,30}$"
        password:
          type: string
          format: password
          description: User password.
          example: testPassword123
          maxLength: 50
          minLength: 8
        email:
          type: string
          format: email
          description: User email address. Must be unique in the system
          example: alex.white@example.com
          pattern: "^[a-z0-9._-]+@[a-z0-9.-]+\\.[a-z]{2,}$"
      required:
      - email
      - password
      - username
    UserDto:
      type: object
      description: User data model returned by the API.
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier in the system
          example: 12345
          readOnly: true
        username:
          type: string
          description: Unique username (login)
          example: alex_white
          readOnly: true
        email:
          type: string
          format: email
          description: User email address
          example: alex.white@example.com
          readOnly: true
    ProblemDetail:
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
  responses:
    ProblemDetailResponse:
      description: RFC 7807 Error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  securitySchemes:
    bearer-token:
      type: http
      description: |
        JWT authentication token obtained from the Authentication Service.
        The API Gateway validates the token and forwards user information in headers:
        - X-User-Id: User identifier
        - X-User-Email: User email address
        - X-User-Username: User username
      scheme: bearer
      bearerFormat: JWT
    internal-service:
      type: apiKey
      description: |
        Internal service authentication for inter-service communication.
        Requires the X-System-Internal-Call header with the origin service name.
        Used for service-to-service calls within the microservices ecosystem.
      name: X-System-Internal-Call
      in: header